move Checkbox to its own component
