+++ internal/ToggleIcon/package.json
